Partial refactor at Dropdown
